pyproject.toml
[project]
name = "servo-controller"
version = "0.1.0"
description = "Servo control system with web interface"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "feetech-servo-sdk>=1.0.0",
    "flask>=3.1.0",
    "numpy>=2.2.3",
    "requests>=2.32.3",
    "pyserial>=3.5",
]

[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["servo_controller", "scripts", "ui"]
